script: Move editing-related code from `DocumentEventHandler` to `editing.rs`
This is preparatory change as we are going to be adding more editing
code here as time goes on. Without this change `DocumentEventHandler` is
around 4000 lines.
