+++ components/script/dom/html/form_controls/htmlinputelement.rs
@@ -2340,11 +2340,8 @@ impl VirtualMethods for HTMLInputElement {
                 .handle_clipboard_event(clipboard_event);
             let flags = reaction.flags;
             if flags.contains(ClipboardEventFlags::FireClipboardChangedEvent) {
-                self.owner_document().event_handler().fire_clipboard_event(
-                    cx,
-                    None,
-                    ClipboardEventType::Change,
-                );
+                self.owner_document()
+                    .fire_clipboard_event(cx, None, ClipboardEventType::Change);
             }
             if flags.contains(ClipboardEventFlags::QueueInputEvent) {
                 self.textinput.borrow().queue_input_event(
